Remove 'game()' call in case of win/loss
Everytime game was called it was causing the click event to fire for the
amount of times game was called. (If scissors was clicked and the game
had restarted twice scissors would play through twice, etc.)

I fixed it by replacing the game call with subtracting both counters
from themselves.